youtube-ai-study/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
youtube-transcript-api==1.2.3
sentence-transformers==2.6.1
reportlab==4.1.0
groq==0.5.0
httpx==0.27.2
numpy==2.2.6
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.32.5
chardet==5.2.0
PyPDF2==3.0.1
Pillow==10.3.0
pytesseract==0.3.10
python-multipart==0.0.9
